.aws-sam/cache/13ad4d5b-9211-4e94-b6de-a56627aeb583/requirements.txt
PyYAML
pandas
requests
scipy
numpy
pytz
python-dateutil
tzdata
